refactor array scoped accumulator ops
