+++ crates/rulemorph/src/transform/operators/array/scoped.rs
@@ -1,6 +1,9 @@
 use super::*;
 
+mod accumulator;
 pub(super) mod predicate;
+
+pub(in crate::transform::operators) use self::accumulator::{eval_array_fold, eval_array_reduce};
 
 pub(in crate::transform::operators) fn eval_array_map(
     args: &[Expr],
@@ -303,104 +306,4 @@ pub(in crate::transform::operators) fn eval_array_sort_by(
 
     let results = items.into_iter().map(|item| item.value).collect::<Vec<_>>();
     Ok(EvalValue::Value(JsonValue::Array(results)))
-}
-
-pub(in crate::transform::operators) fn eval_array_reduce(
-    args: &[Expr],
-    injected: Option<&EvalValue>,
-    record: &JsonValue,
-    context: Option<&JsonValue>,
-    out: &JsonValue,
-    base_path: &str,
-    locals: Option<&EvalLocals<'_>>,
-) -> Result<EvalValue, TransformError> {
-    let total_len = args_len(args, injected);
-    if total_len != 2 {
-        return Err(TransformError::new(
-            TransformErrorKind::ExprError,
-            "expr.args must contain exactly two items",
-        )
-        .with_path(format!("{}.args", base_path)));
-    }
-
-    let array = eval_array_arg(0, args, injected, record, context, out, base_path, locals)?;
-    if array.is_empty() {
-        return Ok(EvalValue::Value(JsonValue::Null));
-    }
-
-    let expr = arg_expr_at(1, args, injected).ok_or_else(|| {
-        TransformError::new(
-            TransformErrorKind::ExprError,
-            "expr.args index is out of bounds",
-        )
-        .with_path(format!("{}.args[1]", base_path))
-    })?;
-    let expr_index = if injected.is_some() { 0 } else { 1 };
-    let expr_path = format!("{}.args[{}]", base_path, expr_index);
-
-    let mut acc = array[0].clone();
-    for (index, item) in array.iter().enumerate().skip(1) {
-        let item_locals = EvalLocals {
-            item: Some(EvalItem { value: item, index }),
-            acc: Some(&acc),
-            pipe: locals.and_then(|locals| locals.pipe),
-            locals: locals.and_then(|locals| locals.locals),
-            precomputed_op_args: locals.and_then(|locals| locals.precomputed_op_args),
-        };
-        let value = eval_expr_or_null(expr, record, context, out, &expr_path, Some(&item_locals))?;
-        acc = value;
-    }
-
-    Ok(EvalValue::Value(acc))
-}
-
-pub(in crate::transform::operators) fn eval_array_fold(
-    args: &[Expr],
-    injected: Option<&EvalValue>,
-    record: &JsonValue,
-    context: Option<&JsonValue>,
-    out: &JsonValue,
-    base_path: &str,
-    locals: Option<&EvalLocals<'_>>,
-) -> Result<EvalValue, TransformError> {
-    let total_len = args_len(args, injected);
-    if total_len != 3 {
-        return Err(TransformError::new(
-            TransformErrorKind::ExprError,
-            "expr.args must contain exactly three items",
-        )
-        .with_path(format!("{}.args", base_path)));
-    }
-
-    let array = eval_array_arg(0, args, injected, record, context, out, base_path, locals)?;
-    let initial =
-        match eval_expr_at_index(1, args, injected, record, context, out, base_path, locals)? {
-            EvalValue::Missing => return Ok(EvalValue::Missing),
-            EvalValue::Value(value) => value,
-        };
-
-    let expr = arg_expr_at(2, args, injected).ok_or_else(|| {
-        TransformError::new(
-            TransformErrorKind::ExprError,
-            "expr.args index is out of bounds",
-        )
-        .with_path(format!("{}.args[2]", base_path))
-    })?;
-    let expr_index = if injected.is_some() { 1 } else { 2 };
-    let expr_path = format!("{}.args[{}]", base_path, expr_index);
-
-    let mut acc = initial;
-    for (index, item) in array.iter().enumerate() {
-        let item_locals = EvalLocals {
-            item: Some(EvalItem { value: item, index }),
-            acc: Some(&acc),
-            pipe: locals.and_then(|locals| locals.pipe),
-            locals: locals.and_then(|locals| locals.locals),
-            precomputed_op_args: locals.and_then(|locals| locals.precomputed_op_args),
-        };
-        let value = eval_expr_or_null(expr, record, context, out, &expr_path, Some(&item_locals))?;
-        acc = value;
-    }
-
-    Ok(EvalValue::Value(acc))
 }
